Update testwlaio.py
The test version of WLAIO should now work and the combiner tab should function properly now on it, the main version the combiner tab does essentially the same thing as the word rules tab. Test will take over the main branch in the near future just gonna test on it some more first.